+++ lz4hc_test.py
@@ -17,7 +17,7 @@ class TestLZ4(unittest.TestCase):
 
     def test_stream_compress(self):
         input_data = "2099023098234882923049823094823094898239230982349081231290381209380981203981209381238901283098908123109238098123"
-        block_size = (len(input_data)/2)+1
+        block_size = int((len(input_data)/2)+1)
 
         stream = lz4hc.create_hc_stream(block_size, lz4hc.COMPRESSIONLEVEL_MAX)
         self.assertNotEqual(stream, None)
